:loud_sound: 联系人

+++ src/main/java/top/felixu/platform/service/manager/ContactorGroupManager.java
@@ -9,13 +9,16 @@ import top.felixu.common.bean.BeanUtils;
 import top.felixu.platform.exception.ErrorCode;
 import top.felixu.platform.exception.PlatformException;
 import top.felixu.platform.model.dto.ContactorTreeDTO;
+import top.felixu.platform.model.entity.Contactor;
 import top.felixu.platform.model.entity.ContactorGroup;
 import top.felixu.platform.model.form.PageRequestForm;
 import top.felixu.platform.service.ContactorGroupService;
+import top.felixu.platform.service.ContactorService;
 
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -28,6 +31,8 @@ import java.util.stream.Collectors;
 public class ContactorGroupManager {
 
     private final ContactorGroupService contactorGroupService;
+
+    private final ContactorService contactorService;
 
     public ContactorGroup getContactorGroupById(Integer id) {
         return contactorGroupService.getContactGroupByIdAndCheck(id);
@@ -51,7 +56,17 @@ public class ContactorGroupManager {
             dto.setDisable(Boolean.TRUE);
             return dto;
         }).collect(Collectors.toList());
-        // TODO: 08/31 填充子节点
+        Map<Integer, List<Contactor>> childrenMap = contactorService.mapByGroupIds(groupIds);
+        result.parallelStream().forEach(group -> {
+            // 没有判断从 Map 中取值的结果不为 null 的原因是 ContactorService#mapByGroupIds 保证了不会为 null
+            group.setChildren(childrenMap.get(group.getKey()).stream().map(contactor -> {
+                ContactorTreeDTO dto = new ContactorTreeDTO();
+                dto.setTitle(contactor.getName());
+                dto.setValue(contactor.getId());
+                dto.setKey(contactor.getId());
+                return dto;
+            }).collect(Collectors.toList()));
+        });
         return result;
     }
 
@@ -69,7 +84,8 @@ public class ContactorGroupManager {
 
     public void delete(Integer id) {
         ContactorGroup group = contactorGroupService.getContactGroupByIdAndCheck(id);
-        // TODO: 08/31 检查是否被联系人使用
+        if (contactorService.countByGroupId(id) > 0)
+            throw new PlatformException(ErrorCode.CONTACTOR_GROUP_USED_BY_CONTACTOR);
         contactorGroupService.delete(group);
     }
